backend/requirements.txt
flask==3.1.1
flask-cors==6.0.1
python-dotenv==1.1.1
supabase==2.17.0
gunicorn==23.0.0
openai==1.99.6